examples/obsidian_integration/server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
dspy-ai
mem0
supermemory
opensearch-py
python-dotenv
PyYAML
